services/poller/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "poller"
version = "0.1.0"
authors = [
  { name="Crow50", email="engineer@example.com" },
]
description = "Rocket Launch Dashboard Poller Service"
requires-python = ">=3.7"
classifiers = [
    "Programming Language :: Python :: 3",
    "Operating System :: OS Independent",
]
dependencies = [
    "SQLAlchemy>=2.0,<3",
    "alembic>=1.7",
    "psycopg[binary]>=3.0",
    "requests",
    "python-dateutil",
    "PyYAML",
    "responses"
]

[project.optional-dependencies]
test = [
    "pytest",
    "pytest-cov"
]

[tool.setuptools.packages.find]
where = ["."]
include = ["poller*"]
